application-service/app/requirements.txt
opencv-python-headless
easyocr
minio
ultralytics
numpy
Pillow
fastapi
python-multipart

